+++ CHANGES.md
@@ -31,6 +31,8 @@ To be released.
     constructor.  [[#666], [#917]]
  -  Added `BigInteger`-typed `previousTotalDifficulty` parameter to
     `Block<T>.Mine()` static method.  [[#666], [#917]]
+ -  Added `options` optional parameter to `Swarm<T>()` constructor.
+    [[#926]]
 
 ### Backward-incompatible network protocol changes
 
@@ -63,6 +65,7 @@ To be released.
  -  Added `InsufficientBalanceException` class.  [[#861], [#900]]
  -  Added `BlockChain<T>.GetBalance()` method.  [[#861], [#900]]
  -  Added `Block<T>.TotalDifficulty` property.  [[#666], [#917]]
+ -  Added `SwarmOptions` class.  [[#926]]
 
 ### Behavioral changes
 
@@ -132,6 +135,7 @@ To be released.
 [#919]: https://github.com/planetarium/libplanet/pull/919
 [#922]: https://github.com/planetarium/libplanet/issues/922
 [#925]: https://github.com/planetarium/libplanet/pull/925
+[#926]: https://github.com/planetarium/libplanet/pull/926
 [sleep mode]: https://en.wikipedia.org/wiki/Sleep_mode